pyproject.toml
[tool.poetry]
name = "genai-voice"
version = "0.1.0"
description = "Generative AI for Voice - A live guided project course."
authors = ["Interview Kickstart"]
readme = "README.md"
packages = [{include = "utils"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "2.2.2"
torchaudio = "2.2.2"
torchvision = "0.17.2"
openai = "1.59.3"
speechrecognition = "3.10.4"
librosa = "0.10.2.post1"
soundfile = "0.12.1"
gtts = "2.5.3"
ffmpeg = "1.4"
ffmpeg-python = "0.2.0"
pydub = "0.25.1"
pyaudio = "0.2.14"
pyttsx3 = "2.98"
gradio = "4.44.1"
ipykernel = "6.29.5"
langchain = "0.3.25"
langchain-core = "0.3.58"
langchain-community = "0.3.1"
langchain-openai = "0.2.8"
langsmith = "0.1.125"
transformers = "4.45.1"
jupyter = "1.1.1"
python-dotenv = "1.0.1"
playwright = "1.47.0"
streamlit = "1.39.0"
streamlit-audiorec = "0.1.3"
watchdog = "5.0.3"
audio-recorder-streamlit = "0.0.10"
ipython = "8.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ExtractWebPagesAndSaveData = "genai_voice.data_utils.extract_web_data:run"
RunChatBotScript = "app.chatbot_gradio_runner:run"
RunChatBotAudioFromFileScript = "app.chatbot_gradio_runner:run_with_file_support"
CallEmmanuelToClass = "test.emmanuel:foo" # poetry install 
